refactor: modify resources-search-config to allow different mappings based on account
