+++ resources-domain/packages/resources-search-config/resourceIndexDocumentMappings/index.ts
@@ -1,0 +1,36 @@
+/**
+ * Copyright (C) 2021-2023 Technology Matters
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see https://www.gnu.org/licenses/.
+ */
+
+import { ResourceIndexDocumentMappings } from './resourceIndexDocumentMappings';
+import khpMappings from './khpMappings';
+
+export * from './resourceIndexDocumentMappings';
+
+export const getMappingsForAccount = (
+  shortCode: string,
+): ResourceIndexDocumentMappings => {
+  switch (shortCode.toUpperCase()) {
+    case 'AS':
+    case 'CA': {
+      return khpMappings;
+    }
+    // TODO
+    case 'USCH':
+    default: {
+      throw new Error(`Mapping not defined for account ${shortCode}`);
+    }
+  }
+};
